vz_chart_helper migrated
