+++ package.json
@@ -78,6 +78,7 @@
     "@types/lodash": "^4.14.149",
     "d3": "^5.14.2",
     "lodash": "^4.17.15",
+    "plottable": "^3.8.6",
     "rxjs": "^6.5.2",
     "zone.js": "^0.9.1"
   }
