+++ tensorboard/components/vz_chart_helpers/BUILD
@@ -1,38 +1,25 @@
-package(default_visibility = ["//visibility:public"])
+load("//tensorboard/defs:defs.bzl", "tf_ts_library")
 
-load("//tensorboard/defs:defs.bzl", "tensorboard_webcomponent_library")
-load("//tensorboard/defs:web.bzl", "tf_web_library")
+package(default_visibility = ["//tensorboard:internal"])
 
 licenses(["notice"])  # Apache 2.0
 
-tf_web_library(
+tf_ts_library(
     name = "vz_chart_helpers",
     srcs = [
-        "plottable-interactions.d.ts",
-        "plottable-interactions.js",
-        "vz-chart-helpers.html",
+        "plottable-interactions.ts",
         "vz-chart-helpers.ts",
-        "vz-chart-tooltip.html",
         "vz-chart-tooltip.ts",
     ],
-    path = "/vz-chart-helpers",
-    deps = [
-        "//tensorboard/components/tf_imports:d3",
-        "//tensorboard/components/tf_imports:lodash",
-        "//tensorboard/components/tf_imports:plottable",
-        "//tensorboard/components/tf_imports:polymer",
-        "//tensorboard/components/vz_sorting",
-    ],
-)
-
-tensorboard_webcomponent_library(
-    name = "legacy",
-    srcs = [":vz_chart_helpers"],
     visibility = ["//visibility:public"],
-    destdir = "vz-chart-helpers",
     deps = [
-        "//tensorboard/components/tf_imports:polymer_lib",
-        "//tensorboard/components/tf_imports_google:lib",
-        "//tensorboard/components/vz_sorting:legacy",
+        "//tensorboard/components/vz_sorting",
+        "@npm//@polymer/decorators",
+        "@npm//@polymer/polymer",
+        "@npm//@types/d3",
+        "@npm//@types/lodash",
+        "@npm//d3",
+        "@npm//lodash",
+        "@npm//plottable",
     ],
 )
