--- tensorboard/components/vz_chart_helpers/vz-chart-helpers.html
@@ -1,24 +1,0 @@
-<!--
-@license
-Copyright 2018 The TensorFlow Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-
-<link rel="import" href="../tf-imports/d3.html" />
-<link rel="import" href="../tf-imports/lodash.html" />
-<link rel="import" href="../tf-imports/plottable.html" />
-<link rel="import" href="../vz-sorting/vz-sorting.html" />
-
-<script src="plottable-interactions.js"></script>
-<script src="vz-chart-helpers.js"></script>
